fix(rem-audit): D1 bind type + silent-error surfacing (CLA-87)
Two bugs in the observability layer surfaced by the first end-to-end
REM run on 2026-05-15:

1. Bind type. worker_rem_audit was casting counter values to i64
   before .into() to JsValue. D1's wasm bindings don't reliably
   round-trip i64 to SQLite INTEGER columns — the bind goes through
   as JS BigInt and the statement silently fails. Existing
   worker_store everywhere uses i32; this module now matches.

2. Silent failures. Audit calls were wrapped in `let _ = ...await`,
   so when bug 1 fired, the errors were dropped on the floor and
   the run summary reported success while no audit rows were
   written. Now: decision-audit failures push into summary.errors;
   run_finish failures (which happen after summary is built) go
   through worker::console_error!.

The actual REM cognitive work was unaffected by either bug —
worker_store primitives kept using i32 binds and wrote correctly,
so the first run produced one real Chopper consolidation in the
memories table and one consolidation_lineage row. Only the
observability surface was broken.

+++ src/worker_rem_audit.rs
@@ -59,14 +59,14 @@ pub async fn record_run_finish(
     )
     .bind(&[
         now.into(),
-        (summary.decayed as i64).into(),
-        (pairs_found as i64).into(),
-        (summary.clusters_attempted as i64).into(),
-        (summary.decisions_created as i64).into(),
-        (summary.decisions_appended as i64).into(),
-        (summary.decisions_revised as i64).into(),
-        (summary.decisions_skipped as i64).into(),
-        (summary.errors.len() as i64).into(),
+        (summary.decayed as i32).into(),
+        (pairs_found as i32).into(),
+        (summary.clusters_attempted as i32).into(),
+        (summary.decisions_created as i32).into(),
+        (summary.decisions_appended as i32).into(),
+        (summary.decisions_revised as i32).into(),
+        (summary.decisions_skipped as i32).into(),
+        (summary.errors.len() as i32).into(),
         errors_json.into(),
         run_id.into(),
     ])?
@@ -104,7 +104,7 @@ pub async fn record_decision(
     .bind(&[
         decision_id.into(),
         run_id.into(),
-        (cluster_idx as i64).into(),
+        (cluster_idx as i32).into(),
         relationship_assessment.into(),
         action.into(),
         members_json.into(),
